Move Mamba2 reference helper out of workloads
